Relocate Kill Detail Alert
Moves the Ganked, Solo, and NPC Alert on the Kill Detail page to the top of the Attackers column. On large kills, this message is hidden off of the page. What is the point of having an alert such as this if nobody can see it? This really only pertains to Ganked Kills such as
https://zkillboard.com/kill/69516063/